some kind of design for admin page, still kinda crappy

+++ client/src/scripts/admin.js
@@ -47,7 +47,9 @@ class AdminPage {
               AdminPage.socket.emit('selectVideo', { _id: this.video._id });
             },
             removeVideo() {
-              this.$http.delete('/video', { body: { _id: this.video._id } });
+              if (window.confirm(`Are you sure you want to delete "${this.video.title ? this.video.title : this.video.url}"`)) {
+                this.$http.delete('/video', { body: { _id: this.video._id } });
+              }
             },
           },
         },
